[CS] Downgrade a Conversion constraint to a Bind
There shouldn't be any @lvalues involved here, as
the two sides of the constraint are a type variable
not marked `TVO_CanBindToLValue`, and a dependent
member for an associated type. Furthermore, the
resulting fixed type will be written out as the
interface type for the tap expr's var, which we
don't want an `@lvalue` for. The constraint system
should introduce that when forming the appropriate
references.