[test] Handle missing availability macros better.
The lit param `availability_macros` appears to be set only from CMake.
Without this commit, when running `lit` manually, `utils/line-directive`
will throw an exception in `expand_response_files` because it is trying
to take the first character of `file_path`, however, this is `''`.

This happens because there is an empty string in the compiler flags,
`-Xfrontend ''` and this is traceable to the `availability_macros`
section of `lit.cfg`. If `availability_macros` is not present in the
params, we return `''`, but `''.split(";")` is `['']`, not `[]`.

Instead, when `availability_macros` is not present, avoid adding the
macro flags entirely. Also, avoid the first-character problem while
we are here as well, since `expand_response_files` is given all arguments
in the command line, and there may conceivably be some need to supply
an empty string argument in the future.