[Diagnostics] A tailored diagnostic for Double<->CGFloat conversion via optional chaining
Implicit conversion between optional types are not currently supported,
this includes optional chaining:

```
struct S {
  var test: CGFloat
}

func compute(s: Double?) -> Double? {
  return 42
}

func test(s: S?) {
  _ = compute(s?.test) // CGFloat? -> Double? implicit conversion
}
```

Since `S` is optional, `test` ends up optional as well and that
wasn't correctly detected, so solver was allowed to form a correct
solution because `CGFloat` is converted to `Double` inside of
the chain _prior_ to it being wrapped into an optional because
contextual type of `Double` is propagated to the result of the
chain earlier that conversion takes place.

Resolves: rdar://83666783