RequirementMachine: Desugar requirements in AbstractGenericSignatureRequest
The requirements passed to this request may have been substituted,
meaning the subject type might be a concrete type and not a type
parameter.

Also, the right hand side of conformance requirements here might be
a protocol composition.

Desugaring converts these kinds of requirements into "proper"
requirements where the subject type is always a type parameter,
which is what the RuleBuilder expects.