SIL: Generalize layout constraints again in subst function type lowering.
To minimize unnecessary abstraction changes, we only want to represent
differences in types that provide interesting representation diversity,
such as between concrete function types with indirect arguments and
generic function types, but not so much between differently-generic
function types. The original subst function type lowering would
handle layout constraints by generalizing them to the most abstract
form that has a similar calling convention, such as
AnyObject -> _RefCountedObject. However, the new implementation
in #40112 moved the generic signature formation for function types
earlier in the type lowering process, and much of type lowering was
unaware of layout constraints other than `requiresClass`, so I
backed off on that optimization to be able to land it. This
adds the optimization back, and generalizes type lowering in
other places to correctly handle other layout constraints.