CanonicalizeInstruction - FIXME and option to preserve debug info
Debug information needs to be preserved at -Onone.

And it should be preserved to a reasonable extend at -O, not
completely thrown away.

This transformation does neither.

The -Onone case is more important. The FIXME explains how that could
be done.

I'm including an example of how to handle the -O case, currently under
a a flag, -sil-load-splitting-debug-info.

This -O support uses debug fragments, which currently crashes when
building SwiftPM's PackageModel module in LLVM's 'X86 Assembly
Printer', llvm::DwarfExpression::addFragmentOffset.

We don't appear to have any verification of debug fragment in either
SIL or LLVM that would catch this. And I haven't been able to reduce
the problem to a small test case.