[CMake] Added support for disabling building the SourceKit XPC service.
Made `BUILD_SOURCEKIT_XPC_SERVICE` into an option so it can be set externally.
It continues to default to `TRUE` if `HAVE_XPC_H AND SWIFT_BUILD_SOURCEKIT` is
true, and `FALSE` otherwise, but this allows a client to disable this ahead of
time and build a sourcekitd.framework without the XPC service.

This addresses <rdar://problem/85511711>.