[DO NOT NERGE] Test, workaround for compiler crash
When trying to compile https://github.com/roberthein/TinyConstraints
with a recent main checkout, there is crash which seems related to the
changes in #38675 and #39664.

The crash was reduced to a test case (included in this PR) in which
a C struct is extended with static function with the same name as one of
the struct fields. If the name is different, the crash does not
reproduce.

The PR also includes a possible fix derived from a check which was added
in #39664 but removed in #38675. Reintroducing the check removes the
problem, and as far as my testing can check, it doesn't create new
problems in the test suite.