[Sema] Made $main non-dynamic.
Previously, it was possible for $main to be made dynamic by passing
-enable-implicit-dynamic.  That has always been a problem in asserts
builds because a direct function_ref to $main is created in the main
function.  More recently, it has become a problem in non-asserts builds
because the dynamic replacements are now in the IR module but are not in
the TBD file.

Here, the function is marked as not dynamic.

rdar://84962693