Enable availability-checking on objc_async.swift test
I'm re-enabling availability checking in the
test/ClangImporter/objc_async.swift test because I want to perform an
availability check in the future.

In re-enabling it, I came across a warning with a fixme and a test
failure on the line. Given that the expected warning has a fixme and is
currently filled in with "something about missing conformance", I think
this is the warning message that they were looking for. Supplying that
gets the test passing with availability-checking enabled.