[cxx-interop] import const reference parameters as __shared T
This change allows Swift code to pass immutable values to const references in C++.
Note that const reference parameters in function template instantiations are not yet
supported in this patch, as even non-const refs are mapped as pointers instead of 'inout' parameters
in Swift. The support for reference parameters in function template instantions will be
added in a follow-up patch.