swift_build_support: Drop support for Python 2
This converts the 'swift_build_support' module to support Python 3
(only). It removes all 'from __future__' imports, conditional python2
imports, and any use of the 'six' module.

It also fixes an issue when attempting to get the length of a 'filter'
result in productpipeline_list_builder.py, using a list comprehension
instead.

It also removes the subprocess workaround for getting output as a
string, instead using the 'text=True' parameter.